requirements.txt
alembic==1.4.3
amqp==2.6.1
billiard==3.5.0.5
celery==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Celery-Helper==1.1.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.2.1
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==2.10.6
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
vine==1.3.0
Werkzeug==1.0.1
zipp==1.2.0
zope.event==4.5.0
zope.interface==5.1.2
